Reorganize after no longer using oh my zsh
I've moved over to using a simple plugin manager (Antidote) and so it no
longer makes sense to structure whole repository according to oh my
zsh's requirements. New structure keeps aliases in top level as a
plugin. Other plugins are grouped in the plugins folder (might change in
the future).